Add support for pmap_enter(..., psind=1) to the i386 pmap.  In other words, add support for explicitly requesting that pmap_enter() create a 2 or 4 MB page mapping.  (Essentially, this feature allows the machine-independent layer to create superpage mappings preemptively, and not wait for automatic promotion to occur.)
Export pmap_ps_enabled() to the machine-independent layer.

Add a flag to pmap_pv_insert_pde() that specifies whether it should fail or
reclaim a PV entry when one is not available.

Refactor pmap_enter_pde() into two functions, one by the same name, that is
a general-purpose function for creating PDE PG_PS mappings, and another,
pmap_enter_4mpage(), that is used to prefault 2 or 4 MB read- and/or
execute-only mappings for execve(2), mmap(2), and shmat(2).

Reviewed by:	kib
Tested by:	pho
Differential Revision:	https://reviews.freebsd.org/D16246


git-svn-id: https://svn.freebsd.org/base/head@336288 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f