config(8): Bump major version after r335998
config-generated hints.c/env.c from r335998 and later are incompatible with
earlier kernels due to no longer setting envmode/hintmode. A minor bump for
this is insufficient, as matching major version with a later minor version
is still viewed as backwards-compatible.

This was an MI kernel change, soo all VERSREQ's are bumped.


git-svn-id: https://svn.freebsd.org/base/head@336353 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f