Make the definition of struct kevent in event.h match what the man page for kevent(2) says.
This is a trivial comment-only fix. The man page for kevent(2) gives
the definition of struct kevent, including a comment on each
field. The actual definition in sys/event.h omitted the comments on
some fields. Add the comments in. Not only does this make the man page
and include file agree, but the comments are useful in and of
themselves.

Reviewed by:	kib (D15778: commented that this should be a separate commit)
MFC after:	3 days
Sponsored by:	Dell EMC


git-svn-id: https://svn.freebsd.org/base/head@336457 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f