MFC: r333766 Add a missing nfsrv_freesession() call for an unlikely failure case.
Since NFSv4.1 clients normally create a single session which supports
both fore and back channels, it is unlikely that a callback will fail
due to a lack of a back channel.
However, if this failure occurred, the session wasn't being dereferenced
and would never be free'd.
Found by inspection during pNFS server development.


git-svn-id: https://svn.freebsd.org/base/stable/11@336516 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f