MFC r336188:
Improve bhyve exit(3) error code.

The bhyve(8) exit status indicates how the VM was terminated:

0	rebooted
1	powered off
2	halted
3	triple fault

The problem is when we have wrappers around bhyve that parses the exit
error code and gets an exit(1) for an error but interprets it as "powered off".
So to mitigate this issue and makes it less error prone for third part
applications, I have added a new exit code 4 that is "exited due to an error".

For now the bhyve(8) exit status are:
0	rebooted
1	powered off
2	halted
3	triple fault
4	exited due to an error

Reviewed by:	@jhb
Sponsored by:	iXsystems Inc.
Differential Revision:	https://reviews.freebsd.org/D16161


git-svn-id: https://svn.freebsd.org/base/stable/11@336704 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f