Re-apply r336625 which was reverted with r336638, now that the underlying pw_scan(3) has been fixed in a way that doesn't perturb other callers of it or the getpwnam(3) family.
Make pw(8) showuser work the same with or without -R <path> for non-root
users.  Without -R, pw(8) uses getpwnam(3), which will open master.passwd
for the root user or passwd for non-root users.  With -R <path> pw(8) was
always opening <path>/master.passwd, which would fail for a non-root user,
then falsely claim the userid you're trying to show doesn't exist.

Now for a non-root user it opens <path>/passwd, and populates the fields in
the returned struct passwd which aren't present in that file with well-known
canonical values, which duplicates the behavior of getpwnam(3).  The net
effect is that the showuser output is identical whether using -R or not.


git-svn-id: https://svn.freebsd.org/base/head@336751 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f