Remove the hopelessly confused GUMSTIX.conf config
This config never worked. At no time did u-boot match the kenrel match
the userland. As all the GUMSTIX gear we support is quite old and/or
not working, remove it. The duovero stuff might work, but nobody
has the hardware for it and GUMSTIX hasn't sold it in years.


git-svn-id: https://svn.freebsd.org/base/head@336776 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f