MFC r307967,324082,325955: config(8) fixes
r307967: Allow config to be compiled from another source directory, such as
one for building tools. This boils down to replacing ${.CURDIR} with
${SRCDIR}, where the latter is the directory in which this makefile
lives.

Also allow overriding where file2c comes from using ${FILE2C}.

r324082: Typo in filename in comment.

r325955: Fix 'local' to not look in the source tree for the file.

Usually 'local' is used along with other rules such as 'no-implicit-rule' or
'dependency' which avoids this problem.  It's possible to need to use
'local' while relying on the default rules though for a file which is not in
the source tree nor generated in the kernel.


git-svn-id: https://svn.freebsd.org/base/stable/11@336938 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f