MFC r336662,r336682
r336662: Deprecate jedec_ts(4) and point users to jedec_dimm(4) instead

jedec_dimm(4) is a superset of the functionality of jedec_ts(4). Mark
jedec_ts(4) as removed in FreeBSD 12, and include a pointer to the migration
instructions in the jedec_dimm(4) manpage, in both the jedec_ts(4) code and
the jedec_ts(4) manpage. Add a note to the jedec_dimm(4) manpage about the
fact that it is a superset of jedec_ts(4).

r336682: Update .Dd in light of r336662.


git-svn-id: https://svn.freebsd.org/base/stable/11@336976 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f