MFC r336363: Process address resolve requests at least one time per second in ibcore.
When setting a large address resolve timeout it was observed that the
address resolving would succeed at the timeout and not when the address
was available. Make sure the address resolving requests are processed no
slower than one time every second.

While at it use "int" for jiffies instead of "unsigned long" to match
FreeBSD ticks.

Sponsored by:		Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@337069 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f