MFC r336366: If the MGID/MLID pair is not on the list return an error in ibcore.
A list of MGID/MLID pairs is built when doing a multicast attach.  When
the multicast detach is called, the list is searched, and regardless of
the search outcome, the driver detach is called.

If an MGID/MLID pair is not on the list, driver detach should not be
called, and an error should be returned.  Calling the driver without
removing an MGID/MLID pair from the list can leave the core and driver
out of sync.

Linux commit:
20c7840a77ddcb2ed2fbd66e8197db2868495751

Sponsored by:		Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@337072 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f