MFC r336386: Honor port_num while resolving GID for IB link layer in ibcore.
ah_attr contains the port number to which cm_id is bound. However, while
searching for GID table for matching GID entry, the port number is
ignored.

This could cause the wrong GID to be used when the ah_attr is converted to
an AH.

Linux commit:
563c4ba3bd2b8b0b21c65669ec2226b1cfa1138b

Sponsored by:		Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@337091 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f